backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.6
semgrep==1.56.0
bandit==1.7.7
openai==1.12.0
anthropic==0.18.0
gitpython==3.1.41
pyyaml==6.0.1
redis==5.0.1
celery==5.3.6
pytest==8.0.0
pytest-asyncio==0.23.4
